return []byte from csv get device data

+++ datafetcher/influx.go
@@ -253,10 +253,10 @@ func (t *TestingInflux) Close() error {
 	orgApi := t.influx.db.OrganizationsAPI()
 	org, err := orgApi.FindOrganizationByName(ctx, TestOrg)
 	if err != nil {
-		if strings.Contains(err.Error(), "not found") {
-			return t.influx.Close()
-		}
 		return fmt.Errorf("finding org: %v", err)
+	}
+	if org == nil {
+		return fmt.Errorf("returned org is nil")
 	}
 	if err = orgApi.DeleteOrganization(ctx, org); err != nil {
 		return fmt.Errorf("deleting org: %v", err)
@@ -275,13 +275,7 @@ func (t *TestingInflux) PrepareDb(allDevicesInfo *deviceinfo.Schema, deviceData 
 	orgApi := t.influx.db.OrganizationsAPI()
 	org, err := orgApi.CreateOrganizationWithName(ctx, TestOrg)
 	if err != nil {
-		if !strings.Contains(err.Error(), "exists") {
-			return fmt.Errorf("org api: %v", err)
-		}
-		org, err = orgApi.FindOrganizationByName(ctx, TestOrg)
-		if err != nil {
-			return fmt.Errorf("getting existing org: %v", err)
-		}
+		return fmt.Errorf("org api: %v", err)
 	}
 	bucketsApi := t.influx.db.BucketsAPI()
 	uniqueNetworks := make([]string, 0, len(allDevicesInfo.DeviceNetworks))
